requirements.txt
# Core ML Libraries
numpy
pandas
scikit-learn

# Data Versioning
dvc
dvc-gdrive

# Experiment Tracking
mlflow

# API Framework
fastapi
uvicorn
pydantic

# Testing
pytest
pytest-cov
# Code Quality
black
flake8
isort
# Utilities
python-dotenv
pyyaml
click
# Visualization
matplotlib
seaborn